+++ lib/poptQV.c
@@ -169,7 +169,7 @@ static void queryArgCallback(poptContext con,
 	if (arg) {
 	    char * qf = (char *)qva->qva_queryFormat;
 	    if (qf) {
-		int len = strlen(qf) + strlen(arg) + 1;
+		size_t len = strlen(qf) + strlen(arg) + 1;
 		qf = xrealloc(qf, len);
 		strcat(qf, arg);
 	    } else {
